canlab_pattern_similarity created, apply_mask update
Tor: I added canlab_pattern_similarity, which takes a matrix of images
(vox x images) and a matrix of pattern masks (one or more, vox x k) and
calculates pattern similarity given your choice of metrics.
Cosine_similarity is looking more and more appealing as a good metric.
Another thing is that if image data are missing voxels, these are
dropped from the calculation. But image data that are not in the
pattern mask are included in the calculation. This does not affect the
dot product, but it affects cosine similarity and correlation metrics,
which is partly why it’s coming up now rather than earlier, i suppose.
This function is now used in apply_mask and extract_roi_averages (which
does pattern expression for a series of local regions).  It should
probably be used in image_similarity_plot, too, but I am hoping Phil
can review all this and implement it there if appropriate.