fmri_data bug fix to deal with recent mask issues-hope ok now?
All, we had a little “destabilization” problem with fmri_data.m
recently. I made a change to make it more flexible on how to handle ‘no
verbose’ input commands and input args, which created a problem with
mask input. Stephan made a change to that but it didn’t handle other
cases right. I tried to fix but didn’t quite get it right. Now I’m
committing a new version, which I hope is more stable.  Do get the
latest repositories before you run, and less to all of us is to be
super careful modifying any of the core functions. We do need to move
to a more systematic unit testing framework before committing anything
to the main repo, but for now we’re still in “development mode” and are
living with these hassles.