Replaced Warning with error.
In setup_X_matrix, replaced a warning about X having no variance with
an error. This warning would previously be ignored, leaving it
ambiguous as to which column was removed, and potentially putting the
wrong column into the output results. Assuming this a problem for some
later function, I decided it would be better to error out than to give
silently incorrect results.