Update print_publication_table.m
Added warnings when the space of the image vector does not match that of the atlas. It could result in more false positives in the output table. It is rooted in the way 'autolabel_regions_using_atlas` works, which calls `resample_space`and `probability_maps_to_region_index` subsequently. 'probability_maps_to_region_index' works in a winner-take-all way such that as long as a voxel is close to a significant voxel, it can become parts of a significant region in the output table. It is not preferrable and should be avoided. Now a heuristic way is to resample the image vector as soon as possible (e.g., before running t tests).